SQL_Alchemy_HW_07/pyproject.toml
[tool.poetry]
name = "sql-alchemy-hw-07"
version = "0.1.0"
description = ""
authors = ["MaxRevan <gladkovnissan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
faker = "^30.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
